# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Banamex CX MVP",
    "version": "0.1.0"
  },
  "paths": {
...
    "/national/compare": {
      "get": {
        "tags": [
          "national"
        ],
        "summary": "National Compare",
        "operationId": "national_compare_national_compare_get",
        "parameters": [
          {
            "name": "month_a",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "YYYY-MM",
              "title": "Month A"
            },
            "description": "YYYY-MM"
          },
          {
            "name": "month_b",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "YYYY-MM",
              "title": "Month B"
            },
            "description": "YYYY-MM"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyComparison"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/national/critical-branches": {
      "get": {
        "tags": [
          "national"
        ],
        "summary": "National Critical Branches",
        "operationId": "national_critical_branches_national_critical_branches_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CriticalBranch"
                  },
                  "title": "Response National Critical Branches National Critical Branches Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/national/causes": {
      "get": {
        "tags": [
          "national"
        ],
        "summary": "National Causes",
        "operationId": "national_causes_national_causes_get",
        "parameters": [
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "Promotor",
                "Pasivo",
                "Detractor"
              ],
              "type": "string",
              "default": "Detractor",
              "title": "Group"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CauseBucket"
                  },
                  "title": "Response National Causes National Causes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/national/strengths": {
      "get": {
        "tags": [
          "national"
        ],
        "summary": "National Strengths",
        "operationId": "national_strengths_national_strengths_get",
        "parameters": [
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "Promotor",
                "Pasivo",
                "Detractor"
              ],
              "type": "string",
              "default": "Promotor",
              "title": "Group"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StrengthBucket"
                  },
                  "title": "Response National Strengths National Strengths Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/national/actions": {
      "get": {
        "tags": [
          "national"
        ],
        "summary": "National Actions",
        "operationId": "national_actions_national_actions_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SuggestedAction"
                  },
                  "title": "Response National Actions National Actions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CauseBucket": {
        "properties": {
          "bucket": {
            "type": "string",
            "title": "Bucket"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "pct_of_group": {
            "type": "number",
            "title": "Pct Of Group"
          },
          "sample_l2": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sample L2"
          }
        },
        "type": "object",
        "required": [
          "bucket",
          "count",
          "pct_of_group",
          "sample_l2"
        ],
        "title": "CauseBucket"
      },
...
      "CriticalBranch": {
        "properties": {
          "branch_id": {
            "type": "string",
            "title": "Branch Id"
          },
          "nps_actual": {
            "type": "number",
            "title": "Nps Actual"
          },
          "nps_target": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nps Target"
          },
          "gap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gap"
          },
          "detractors_pct": {
            "type": "number",
            "title": "Detractors Pct"
          },
          "triggered_conditions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Triggered Conditions"
          }
        },
        "type": "object",
        "required": [
          "branch_id",
          "nps_actual",
          "nps_target",
          "gap",
          "detractors_pct",
          "triggered_conditions"
        ],
        "title": "CriticalBranch"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MonthlyComparison": {
        "properties": {
          "month_a": {
            "type": "string",
            "title": "Month A"
          },
          "month_b": {
            "type": "string",
            "title": "Month B"
          },
          "nps_a": {
            "type": "number",
            "title": "Nps A"
          },
          "nps_b": {
            "type": "number",
            "title": "Nps B"
          },
          "nps_change": {
            "type": "number",
            "title": "Nps Change"
          },
          "distribution_a": {
            "$ref": "#/components/schemas/NPSDistribution"
          },
          "distribution_b": {
            "$ref": "#/components/schemas/NPSDistribution"
          },
          "causes_a": {
            "items": {
              "$ref": "#/components/schemas/CauseBucket"
            },
            "type": "array",
            "title": "Causes A"
          },
          "causes_b": {
            "items": {
              "$ref": "#/components/schemas/CauseBucket"
            },
            "type": "array",
            "title": "Causes B"
          },
          "causes_increased": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Causes Increased"
          },
          "causes_decreased": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Causes Decreased"
          },
          "strengths_a": {
            "items": {
              "$ref": "#/components/schemas/StrengthBucket"
            },
            "type": "array",
            "title": "Strengths A"
          },
          "strengths_b": {
            "items": {
              "$ref": "#/components/schemas/StrengthBucket"
            },
            "type": "array",
            "title": "Strengths B"
          },
          "strengths_increased": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths Increased"
          },
          "strengths_decreased": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths Decreased"
          },
          "branches_improved": {
            "items": {
              "$ref": "#/components/schemas/CriticalBranch"
            },
            "type": "array",
            "title": "Branches Improved"
          },
          "branches_worsened": {
            "items": {
              "$ref": "#/components/schemas/CriticalBranch"
            },
            "type": "array",
            "title": "Branches Worsened"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/SuggestedAction"
            },
            "type": "array",
            "title": "Actions"
          }
        },
        "type": "object",
        "required": [
          "month_a",
          "month_b",
          "nps_a",
          "nps_b",
          "nps_change",
          "distribution_a",
          "distribution_b",
          "causes_a",
          "causes_b",
          "causes_increased",
          "causes_decreased",
          "strengths_a",
          "strengths_b",
          "strengths_increased",
          "strengths_decreased",
          "branches_improved",
          "branches_worsened",
          "actions"
        ],
        "title": "MonthlyComparison"
      },
...
      "NPSDistribution": {
        "properties": {
          "promoters_pct": {
            "type": "number",
            "title": "Promoters Pct"
          },
          "passives_pct": {
            "type": "number",
            "title": "Passives Pct"
          },
          "detractors_pct": {
            "type": "number",
            "title": "Detractors Pct"
          },
          "promoters_count": {
            "type": "integer",
            "title": "Promoters Count"
          },
          "passives_count": {
            "type": "integer",
            "title": "Passives Count"
          },
          "detractors_count": {
            "type": "integer",
            "title": "Detractors Count"
          }
        },
        "type": "object",
        "required": [
          "promoters_pct",
          "passives_pct",
          "detractors_pct",
          "promoters_count",
          "passives_count",
          "detractors_count"
        ],
        "title": "NPSDistribution"
      },
...
      "StrengthBucket": {
        "properties": {
          "bucket": {
            "type": "string",
            "title": "Bucket"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "pct_of_group": {
            "type": "number",
            "title": "Pct Of Group"
          },
          "sample_l2": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sample L2"
          }
        },
        "type": "object",
        "required": [
          "bucket",
          "count",
          "pct_of_group",
          "sample_l2"
        ],
        "title": "StrengthBucket"
      },
      "SuggestedAction": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "priority": {
            "type": "string",
            "enum": [
              "alta",
              "media",
              "baja"
            ],
            "title": "Priority"
          },
          "related_bucket": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Related Bucket"
          },
          "related_branches": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Related Branches",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "text",
          "priority",
          "related_bucket"
        ],
        "title": "SuggestedAction"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}